server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
torch>=2.0.0
ultralytics>=8.0.0
imagehash>=4.3.1
aiosqlite>=0.19.0
pyyaml>=6.0
httpx>=0.25.0